rename verifier zkapp file
